grpc_backend/images_service/pyproject.toml
[project]
name = "images-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio",
    "pyjwt[crypto]",
    "grpcio-tools"
]
